Move Settings to lib/ so we can use it in SetupPage
